% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HPD.R
\name{binary2ternary}
\alias{binary2ternary}
\title{Transform 2nd order barycentric coordinates to 3rd order}
\usage{
binary2ternary(binary_points, ternary_cols)
}
\arguments{
\item{binary_points}{matrix of points, with each row a 2-vector on the 1-simplex}

\item{ternary_cols}{2-vector specifying the columns of the result
corresponding to the two columns of \code{binary_points}}
}
\value{
matrix of points, with each row a 3-vector on the 2-simplex. The first
element of \code{ternary_cols} gives the column of the result that takes the values
of the first column of \code{binary_points}; the second element gives the column of
the result taking the values of the second column of \code{binary_points}.
The remaining column of the result is set to zero.
}
\description{
\code{binary2ternary} transforms the 2nd order barycentric coordinates of the
rows of \code{binary_points} into the 3rd order barycentric coordinates of the
rows of \code{ternary_points}.
This is useful for plotting points and line segments of binary choice
probabilities on the sides of the 3rd order barycentric coordinate system.
The two non-zero columns of the result are specified in the 2-vector
ternary_cols.
}
\examples{
binary2ternary(matrix(c(0.6, 0.4), nrow=1), c(2,3)) # returns 1x3 matrix [0 0.6 0.4]
binary2ternary(matrix(c(0.6, 0.4), nrow=1), c(3,2)) # returns 1x3 matrix [0 0.4 0.6]
}
