% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BM_terms.R
\name{BM_terms}
\alias{BM_terms}
\title{Compute Block-Marschack terms from a random choice structure}
\usage{
BM_terms(P)
}
\arguments{
\item{P}{A random choice structure. Element \code{P[A,x]} is the probability
that \eqn{x} is chosen when choice set \eqn{A} is presented.}
}
\value{
A matrix of Block-Marschak terms.
If \code{Q} is the result, then for all random rankings of the elements of the
universe of objects than induce the random choice structure \eqn{P},
element \code{Q[A,x]} is the probability that \eqn{x} is ranked higher than
any other element of choice set \eqn{A} and lower than any element in the
complement of \eqn{A} in the universe of objects.
If \eqn{x \notin A}, then \code{Q[A,x]} has value \code{NA}.
The RCS \eqn{P} satisfies random utility if and only if for all
\eqn{x \in A \subseteq T}, \eqn{Q_A(x) \geq 0}
}
\description{
\code{BM_terms} returns a matrix of the same dimensions as the input matrix, a
random choice structure \code{P}.
The non-NA terms are all non-negative if and only if the random choice structure
\code{P} satisfies random utility.
}
\examples{
P <- create_P3(0.7, 0.6, 0.8, 0.6, 0.3, names = c('x', 'y', 'z'))
Q <- BM_terms(P)
Q
}
