% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compromise.R
\name{compromise}
\alias{compromise}
\title{Check if choice probabilities exhibit compromise effect}
\usage{
compromise(P, target, competitor, decoy, two_sided = FALSE)
}
\arguments{
\item{P}{matrix, random choice structure}

\item{target}{index of target object}

\item{competitor}{index of competitor object}

\item{decoy}{index of decoy object}

\item{two_sided}{logical value indicating whether the compromise inequality
is to hold also for competitor and decoy exchanged.}
}
\value{
A logical value indicating whether the appropriate choice probabilities are
in the similarity effect region.
}
\description{
\code{compromise} returns TRUE if the specified binary and ternary
choice probabilities in random choice structure \code{P} are in the specified
compromise effect region.
}
\examples{
P <- create_P3(0.7, 0.6, 0.8, 0.6, 0.3, names = c('x', 'y', 'z'))
compromise(P, target=1, competitor=2, decoy=3)

}
