% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_P3.R
\name{create_P3}
\alias{create_P3}
\title{Random Choice Structure for a three-object universe}
\usage{
create_P3(p12, p23, p13, P1, P2, names = c("x", "y", "z"))
}
\arguments{
\item{p12}{Probability of chosing object 1 when presented with objects 1 and 2}

\item{p23}{Probability of chosing object 2 when presented with objects 2 and 3}

\item{p13}{Probability of chosing object 1 when presented with objects 1 and 3}

\item{P1}{Probability of chosing object 1 when presented with objects 1, 2 and 3}

\item{P2}{Probability of chosing object 2 when presented with objects 1, 2 and 3}

\item{names}{character vector giving names to the three objects}
}
\value{
A Random Choice Structure
}
\description{
\code{create_P3} creates a random choice structure for a three-object universe
from
}
\examples{
P <- create_P3(21/40, 37/40, 28/40, 19/40, 15/40, names=c('Red', 'Purple', 'Pink'))
P
}
