% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HPD.R
\name{dDir3_quantile}
\alias{dDir3_quantile}
\title{Quantile of third order Dirichlet density value}
\usage{
dDir3_quantile(quantile, alpha, normalized = FALSE)
}
\arguments{
\item{quantile}{the quantile of the desired density value}

\item{alpha}{a vector of Dirichlet parameters}

\item{normalized}{binary; if TRUE, return the quantile as a fraction of
the maximum density value; if FALSE, return the unnormalized quantile.}
}
\value{
The value of the quantile, normalized or not
}
\description{
\code{dDir3_quantile} computes an approximation of the given
quantile of a third order Dirichlet density value, under that Dirichlet
distribution.
}
