% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HPD.R
\name{dDir_max}
\alias{dDir_max}
\title{Maximum density of a Dirichlet distribution}
\usage{
dDir_max(alpha, log = TRUE)
}
\arguments{
\item{alpha}{vector of Dirichlet parameters.}

\item{log}{logical; if TRUE, the log maximum density is returned.}
}
\value{
Density or log density value.
}
\description{
\code{max_dDir} computes the maximum density of a Dirichlet distribution
as a function of the parameter vector alpha.
}
