% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HPD.R
\name{dDir_moments}
\alias{dDir_moments}
\title{Moments of Dirichlet density values}
\usage{
dDir_moments(beta, n_mu, log = FALSE)
}
\arguments{
\item{n_mu}{number of moments to compute.}

\item{log}{logical; if true return log moments.}

\item{alpha}{vector of Dirichlet parameters.}
}
\value{
vector of moments
}
\description{
\code{moments_dDi} computes a vector of the first n raw moments of Dirichlet
density values, under that Dirichlet distribution.
}
