% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HPD.R
\name{Dir3_HD_region}
\alias{Dir3_HD_region}
\title{Compute highest density (HD) region for a third order Dirichlet distribution}
\usage{
Dir3_HD_region(alpha, HD_probability)
}
\arguments{
\item{alpha}{a vector of three (positive) Dirichlet parameters.}

\item{HD_probability}{probability of region to construct}
}
\value{
polygon approximation of HD region.
}
\description{
This function computes a polygon approximating the highest density region
of a third order Dirichlet distribution. This can be used to compute highest
prior density and highest posterior density (HPD) regions.
}
