% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HPD.R
\name{Dir_mult_ML}
\alias{Dir_mult_ML}
\title{Marginal likelihood for Dirichlet-multinomial model}
\usage{
Dir_mult_ML(alpha, N, log = TRUE)
}
\arguments{
\item{alpha}{vector of Dirichlet parameters}

\item{N}{vector of multinomial counts}

\item{log}{logical; if TRUE, return the log Bayes factor.}
}
\value{
Marginal likelihood or log marginal likelihood
}
\description{
\code{Dir_mult_ML} computes the marginal likelihood for a Dirichlet prior
and multinomial data generating process.
}
