% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HPD.R
\name{Ind_Dir_mult_ML}
\alias{Ind_Dir_mult_ML}
\title{Marginal likelihood for independent Dirichlet-multinomial model}
\usage{
Ind_Dir_mult_ML(A, N, log = TRUE)
}
\arguments{
\item{A}{matrix of Dirichlet parameters, each row giving the Dirichlet
distribution of the corresponding row of a random choice structure.}

\item{N}{count matrix for a universe of objects.}

\item{log}{logical; if TRUE, return the log Bayes factor}
}
\description{
\code{Ind_Dir_mult_ML} computes the marginal likelihood for a model
where rows of a count matrix are independent multinomial and the
rows of the unknown random choice structure are a priori independent
Dirichlet.
}
