% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{log_L_Dir_mult}
\alias{log_L_Dir_mult}
\title{Log likelihood, Dirichlet-multinomial model}
\usage{
log_L_Dir_mult(alpha, n, categorical = FALSE, log = TRUE)
}
\arguments{
\item{alpha}{vector of Dirichlet parameters}

\item{n}{vector of multinomial counts}

\item{categorical}{logical; if \code{TRUE}, the likelihood is the for
the sequence of responses (categorical distribution) rather than for
the counts (multinomial distribution).}

\item{log}{logical; if \code{TRUE}, return the log likelihood;
if \code{FALSE}, the likelihood.
\code{log=FALSE} is usually not recommendend, as underflow is likely.}
}
\value{
Likelihood or log likelihood
}
\description{
\code{log_L_Dir_mult} computes the log marginal likelihood for a multinomial
data generating process and a Dirichlet prior over choice probabilities.
}
\examples{
log_L_Dir_mult(c(2.4, 1.5, 3.2), c(45, 20, 33))
}
