% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{log_L_multinomial}
\alias{log_L_multinomial}
\title{Log likelihood for multinomial model}
\usage{
log_L_multinomial(p, n, log = TRUE)
}
\arguments{
\item{p}{vector of probabilities}

\item{n}{vector of counts}

\item{log}{logical; if \code{TRUE}, return the log likelihood;
if \code{FALSE}, the likelihood.
\code{log=FALSE} is usually not recommendend, as underflow is likely.}
}
\value{
log likelihood or likelihood
}
\description{
\code{log_L_multinomial} computes the log likelihood function for a
multinomial model
}
