% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{log_ML_DCE_vector_alpha_Dir_mult}
\alias{log_ML_DCE_vector_alpha_Dir_mult}
\title{Marginal likelihood for DCE, vector Dirichlet multinomial model}
\usage{
log_ML_DCE_vector_alpha_Dir_mult(u, N, log = TRUE)
}
\arguments{
\item{u}{vector determining Dirichlet parameters}

\item{N}{count matrix}

\item{log}{logical; if \code{TRUE}, return the log marginal likelihood;
if \code{FALSE}, the marginal likelihood.
\code{log=FALSE} is usually not recommendend, as underflow is likely.}
}
\description{
\code{log_ML_DCE_vector_alpha_Dir_mult} computes the marginal likelihood for a model
where choice probability vectors \eqn{P_A} are mutually independent, and
\deqn{P_A(x_1,\ldots,x_{|A|}) \sim \mathrm{Di}(u_1,\ldots,u_|A|)},
and choices are conditionally independent multinomial.
}
