% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HPD.R
\name{log_ML_Dir_mult}
\alias{log_ML_Dir_mult}
\title{Log marginal likelihood for Dirichlet-multinomial model}
\usage{
log_ML_Dir_mult(alpha, n, log = TRUE)
}
\arguments{
\item{alpha}{vector of Dirichlet parameters}

\item{n}{vector of multinomial counts}

\item{log}{logical; if \code{TRUE}, return the log marginal likelihood;
if \code{FALSE}, the marginal likelihood.
\code{log=FALSE} is usually not recommendend, as underflow is likely.}
}
\value{
Marginal likelihood or log marginal likelihood
}
\description{
\code{log_ML_Dir_mult} computes the log marginal likelihood for a multinomial
data generating process and a Dirichlet prior over choice probabilities.
}
