% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MC_data.R
\docType{data}
\name{MC_counts}
\alias{MC_counts}
\title{Choice counts in multiple choice experiment}
\format{An object of class \code{array} of dimension 141 x 31 x 5.}
\usage{
MC_counts
}
\description{
A 141 by 31 by 5 array of count data.
Element \eqn{i,j,k} gives the number of times subject \eqn{i} chose object \eqn{k} when
presented with choice set \eqn{j}.
The choice set index j=1,...,31 encodes a non-empty subset of the universe
of the five choice objects a, b, c, d and e, numbered 1, 2, 3, 4, 5.
Whenever \eqn{k} is not an element of \eqn{j}, the value is \code{NA}.
}
\seealso{
\code{\link{RanCh}}, under \code{Datasets} for a description of the experiment.
Other data objects for the experiment include
\code{\link{MC_trials}}, a table of choice trial data, and
\code{\link{MC_raw}}, a table of raw data.
\code{\link{set_index}} describes how the choice set index encodes the choice set.
}
\keyword{datasets}
