% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MC_data.R
\docType{data}
\name{MC_raw}
\alias{MC_raw}
\title{Multiple Choice experiment data}
\format{A data frame with 53 variables generated in (WJM)
We do not provide documentation, although the variable names are suggestive.}
\usage{
MC_raw
}
\description{
Raw trial-by-trial data from Multiple Choice experiment
}
\seealso{
\code{\link{RanCh}}, under \code{Datasets} for a description of the experiment.
Other data objects for the experiment include
\code{\link{MC_trials}}, a table of choice trial data, and
\code{\link{MC_counts}}, an array of choice count data.
}
\keyword{datasets}
