% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MC_data.R
\docType{data}
\name{MC_trials}
\alias{MC_trials}
\title{Table of choice trial data from the Multiple Choice experiment}
\format{A tibble with 18 variables
\describe{
\item{\code{subject}}{subject identifier}
\item{\code{trial}}{trial identifier (gives the order in which a subject sees choice sets)}
\item{\code{set}}{factor, name of choice set presented: 'ab', 'cde', etc., with objects
in alphabetical order}
\item{\code{choice}}{factor, choice made by subject: 'a', 'b', 'c', 'd' or 'e'}
\item{\code{set_perm}}{factor, order of presentation of objects on screen, left to right}
\item{\code{set_card}}{integer, cardinality of choice set (i.e. number of available options)}
\item{\code{set_bin}}{binary representation of choice set (binary digits indicate object membership in choice set)}
\item{\code{choice_int}}{integer, code for chosen object: a=1, b=2, ..., e=5}
\item{\code{ab}, \code{ac}, \ldots, \code{de}}{revealed preference indicator:
taking column ab as an example, the value is 1 if a is revealed preferred to b,
-1 if b is revealed preferred to a, 0 otherwise.}}}
\usage{
MC_trials
}
\description{
Table of choice trial data from the Multiple Choice experiment
}
\seealso{
\code{\link{RanCh}}, under \code{Datasets} for a description of the experiment.
Other data objects for the experiment include
\code{\link{MC_counts}}, an array of choice count data, and
\code{\link{MC_raw}}, a table of raw data.
}
\keyword{datasets}
