% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplicative_X3.R
\name{multiplicative_X3}
\alias{multiplicative_X3}
\title{Compute a cross section of the multiplicative inequality region}
\usage{
multiplicative_X3(P)
}
\arguments{
\item{P}{A random choice structure}
}
\value{
A 3x3 matrix where each row gives one of the three vertices, in
barycentric coordinates, of the triangular region where the multiplicative
inequality holds.
}
\description{
\code{multiplicative_X3} computes the region (a triangle) of ternary
probabilities consistent with given binary probabilities and the
multiplicative inequality.
}
\examples{
P = create_P3(0.7, 0.6, 0.8, 0.6, 0.3, 0.1, names = c('x', 'y', 'z'))
multiplicative_X3(P)
}
