% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{P_Luce}
\alias{P_Luce}
\title{P_Luce}
\usage{
P_Luce(v)
}
\arguments{
\item{v}{postive vector, Luce weights}
}
\value{
RCS with choice probabilities given by Luce weights
}
\description{
\code{P_Luce} constructs a random choice structure (RCS) from a Luce
model.
}
\examples{
P = P_Luce(c(1.2,2.5,3.3))
}
