% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{P_uniform}
\alias{P_uniform}
\title{P_uniform}
\usage{
P_uniform(n_objects)
}
\arguments{
\item{n_objects}{integer, number of objects in universe}
}
\value{
RCS where all choice probabilities are discrete uniform
}
\description{
\code{P_uniform} constructs a random choice structure (RCS) where all
choice probabilities are discrete uniform
}
\examples{
P <- P_uniform(3)
}
