% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PC_data.R
\docType{data}
\name{PC_counts}
\alias{PC_counts}
\title{Array of choice count data from the Population Choice experiment}
\format{An object of class \code{array} of dimension 32 x 31 x 5.}
\usage{
PC_counts
}
\description{
A 32 by 31 by 5 array of count data.
Element \eqn{i,j,k} gives the number of subjects that chose object \eqn{k} when
presented with choice set \eqn{j} of domain \eqn{i}.
The choice set index j=1,...,31 encodes a non-empty subset of the universe
of the five choice objects a, b, c, d and e, numbered 1, 2, 3, 4, 5.
Whenever \eqn{k} is not an element of \eqn{j}, the value is \code{NA}.
}
\seealso{
{\code{\link{RanCh}}}, under \code{Datasets} for a description of the experiment.
Other data objects for the experiment include
\code{\link{PC_trials}}, a table of choice trial data,
\code{\link{PC_demographics}}, a table of demographic information, and
\code{\link{PC_raw}}, a table of raw data.
\code{\link{set_index}} describes how the choice set index encodes the choice set.
}
\keyword{datasets}
