% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PC_data.R
\docType{data}
\name{PC_raw}
\alias{PC_raw}
\title{Population Choice experiment data}
\format{A data frame with 17 variables:
\describe{
\item{\code{design}}{}
\item{\code{gender}}{Sex of respondant: 1 for male, 2 for female}
}}
\usage{
PC_raw
}
\description{
Record of every choice made by every respondant.
}
\keyword{datasets}
