% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PC_data.R
\docType{data}
\name{PC_trials}
\alias{PC_trials}
\title{Record of all choice trials}
\format{A data frame with 14 variables
\describe{
\item{\code{subj}}{Subject identifier}
\item{\code{domain}}{Factor indicating choice domain}
\item{\code{trial}}{Trial identifier (gives the order in which a subject sees choice sets)}
\item{\code{subs}}{Factor indicating the choice subset presented: 'ab', 'cde', etc., objects
always in alphabetical order}
\item{\code{choice}}{Factor indicating the choice made: 'a', 'b', 'c', 'd' or 'e'}
\item{\code{subs_conf}}{Subset configuration, the order objects appear on the screen: 'ba', 'ecd', etc.,
objects not necessarily in alphabetical order}
\item{\code{subs_bin}}{Code for subset where digits of binary representation indicate object membership}
\item{\code{choice_int}}{Integer code for chosen object}
\item{\code{ab}}{Revealed preference indicator: 1 for a revealed preferred to b,
-1 for b revealed preferred to a, 0 otherwise.
This is the first of ten revealed preference columns, each pertaining to a particular
doubleton set.}
}}
\usage{
PC_trials
}
\description{
Record of all choice trials
}
\keyword{datasets}
