% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_P3.R
\name{plot_P3}
\alias{plot_P3}
\title{Plot a Random Choice Structure in barycentric coordinates}
\usage{
plot_P3(P, perm = c(1, 2, 3), binary_pch = 1, ternary_pch = 20)
}
\arguments{
\item{P}{A random choice structure for a universe of three objects}

\item{perm}{A permutation of (1, 2, 3) specifying which objects in the universe
correspond to the bottom left, top, and bottom right vertex, respectively of
the ternary plot.}

\item{binary_pch}{Plotting character (pch) for binary choice probabilities. Defaults
to a hollow circle.}

\item{ternary_pch}{Plotting character (pch) for ternary choice probability. Defaults
to a solid circle. The convention established with the defaults for binary_pch and
ternary_pch allow one to distinguish between a binary choice probability and a ternary
choice probability that happens to be on the boundary of the triangle.}
}
\description{
\code{plot_P3} plots four points specifying a Random Choice Structure for
a universe of three objects.
}
\examples{
P = create_P(0.7, 0.6, 0.8, 0.6, 0.3, 0.1, names = c('x', 'y', 'z'))
plot_P3(P)
}
