% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HPD.R
\name{prior_DCE_scalar_alpha}
\alias{prior_DCE_scalar_alpha}
\title{One-parameter Dirichlet prior for a RCS}
\usage{
prior_DCE_scalar_alpha(alpha, n_objects)
}
\arguments{
\item{alpha}{univariate parameter for the one-parameter Dirichlet prior.}

\item{n_objects}{number of objects in the universe.}
}
\value{
a matrix of Dirichlet parameters with the same dimensions as a
count matrix for a universe of the same size.
}
\description{
\code{prior_DCE_scalar_alpha} computes a matrix of Dirichlet parameters for a
one-parameter Dirichlet prior for a random choice structure.
}
