% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportions.R
\name{proportions}
\alias{proportions}
\title{Random Choice Structure from count proportions}
\usage{
proportions(N)
}
\arguments{
\item{N}{A count matrix.}
}
\value{
A random choice structure.
}
\description{
\code{proportions} takes a count matrix as input, and returns choice
proportions as a random choice structure.
}
\examples{
PC_P = proportions(PC_counts)
}
