% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BM_terms.R
\name{random_utility}
\alias{random_utility}
\title{Check if random choice structure satisfies random utility}
\usage{
random_utility(P)
}
\arguments{
\item{P}{A random choice structure}
}
\value{
A logical value indicating whether \code{P} satisfies random utility.
}
\description{
\code{random_utility} returns TRUE if the random choice structure P satisfies
random utility and FALSE otherwise.
}
\examples{
P <- create_P3(0.7, 0.6, 0.8, 0.6, 0.3, names = c('x', 'y', 'z'))
random_utility(P)

}
