% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{RCS_scalar_alpha_prior}
\alias{RCS_scalar_alpha_prior}
\title{Scalar alpha prior for a RCS}
\usage{
RCS_scalar_alpha_prior(alpha, n_objects)
}
\arguments{
\item{alpha}{univariate parameter for the one-parameter Dirichlet prior.}

\item{n_objects}{number of objects in the universe.}
}
\value{
a matrix of Dirichlet parameters for the scalar alpha prior
}
\description{
\code{RCS_scalar_alpha_prior} constructs a matrix of Dirichlet parameters
for a one-parameter Dirichlet prior for a random choice structure.
}
\examples{
Alpha <- RCS_scalar_alpha_prior(2.5, 3)
P <- sim_RCS_Dirichlet(10, Alpha)
}
