% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HPD.R
\name{rDir}
\alias{rDir}
\title{Dirichlet random variates}
\usage{
rDir(n, alpha)
}
\arguments{
\item{n}{number of draws}

\item{alpha}{vector of Dirichlet parameters}
}
\value{
matrix with n rows, each a draw from the Dirichlet distribution
}
\description{
\code{rDir} draws from the Dirichlet distribution
}
\examples{
library(klaR)
p <- rDir(1000, c(1, 1, 1)) # Uniform distribution on 2-simplex
triplot(label=c('x', 'y', 'z'))
plot(tritrafo(p))
}
