% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regularity.R
\name{regularity}
\alias{regularity}
\title{Check if random choice structure satisfies regularity}
\usage{
regularity(P)
}
\arguments{
\item{P}{A random choice structure}
}
\value{
A logical value indicating whether \code{P} satisfies regularity.
}
\description{
\code{regularity} returns \code{TRUE} if the random choice structure \code{P} satisfies
regularity and \code{FALSE} otherwise.
}
\examples{
P <- create_P3(0.7, 0.6, 0.8, 0.6, 0.3, names = c('x', 'y', 'z'))
regularity(P)

}
