% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regularity_X3.R
\name{regularity_X3}
\alias{regularity_X3}
\title{Compute a cross section of the regularity region}
\usage{
regularity_X3(P)
}
\arguments{
\item{P}{A random choice structure.}
}
\value{
If the region is empty, the output is NULL. Otherwise, a 3x3 matrix
  where each row gives one of the three vertices in barycentric coordinates.
}
\description{
\code{regularity_X3} computes the region (a triangle or the empty set) of
ternary probabilities consistent with given binary probabilities and the
regularity condition.
}
\examples{
P = create_P(0.7, 0.6, 0.8, 0.6, 0.3, 0.1, names = c('x', 'y', 'z'))
reg_region = regularity_X3(P)
}
