% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BM_terms.R
\name{ru}
\alias{ru}
\title{Helper function for random_utility and BM_terms, below.}
\usage{
ru(P, compute_Q)
}
\arguments{
\item{P}{matrix, a random choice structure}

\item{compute_Q, }{a logical value indicating whether or not to compute
\eqn{Q}, the matrix of Block-Marschak terms corresponding to \code{P}.
If compute_Q is true, output is a list with elements Q and is_ru, where is_ru is a
boolean value that is TRUE or FALSE according to whether or not P satisfies random utility
If compute_Q is FALSE, output is TRUE or FALSE according to whether or not P satisfies
random utility.}
}
\value{
A logical value indicating whether \code{P} satisfies random utility.
}
\description{
\code{ru} tests whether or not a random choice structure satisfies random utility
}
