% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{similarity}
\alias{similarity}
\title{Check if choice probabilities exhibit similarity effect}
\usage{
similarity(P, target, competitor, decoy, two_sided = FALSE)
}
\arguments{
\item{P}{matrix, random choice structure}

\item{target}{index of target object}

\item{competitor}{index of competitor object}

\item{decoy}{index of decoy object}

\item{two_sided}{logical value indicating whether the similarity inequality
is to hold also for target and decoy exchanged.}
}
\value{
A logical value indicating whether the appropriate choice
probabilities are in the similarity effect region.
}
\description{
\code{similarity} returns TRUE if the specified binary and ternary
choice probabilities in random choice structure \code{P} are in the
specified similarity effect region.
}
\examples{
P <- create_P3(0.7, 0.6, 0.8, 0.6, 0.3, names = c('x', 'y', 'z'))
similarity(P, target=1, competitor=2, decoy=3)

}
