% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity_X3.R
\name{similarity_X3}
\alias{similarity_X3}
\title{Similarity effect regions}
\usage{
similarity_X3(pxz, pyz)
}
\arguments{
\item{pxz}{scalar, probability of choosing \eqn{y} when presented
with choice set \eqn{\{x,z\}}}

\item{pyz}{scalar, probability of choosing \eqn{y} when presented
with choice set \eqn{\{y,z\}}}
}
\value{
named list of six polygons in a barycentric coordinate system.
Each polygon is a matrix, with one row for each polygon vertex and one column
for each of the three choice objects \eqn{x}, \eqn{y} and \eqn{z}.
Element \eqn{i,j} gives the probability of choosing \eqn{j}, when presented with choice
set \eqn{\{x,y,z\}}, at the \eqn{i}'th polygon vertex.
The six polygons---triangles except for \code{Sxy}, a quadrilateral---are:
\describe{
\item{\code{Sxyz}}{region where there is a similarity effect with object \eqn{x} as target,
\eqn{y} as decoy and \eqn{z} as competitor}
\item{\code{Syxz}}{region where there is a similarity effect with object \eqn{x} as target,
\eqn{y} as decoy and \eqn{z} as competitor}
\item{\code{So}}{region where there is neither similarity effect}
\item{\code{Sx}}{region with only the \eqn{x} effect}
\item{\code{Sy}}{region with only the \eqn{y} effect}
\item{\code{Sxy}}{region with both effects}
}
}
\description{
\code{similarity_X3} constructs six regions (all polygons) associated with
binary-ternary similary effects.
Below, \eqn{x} and \eqn{y} are "similar" choice objects and object \eqn{z} is dissimilar to both
\eqn{x} and \eqn{y}.
The interior of each polygon gives the set of ternary choice probabilities
consistent with (1) the two specified binary choice probabilities and (2) one of six
similarity effect conditions.
}
\examples{
S <- similarity_X3(0.5, 0.6)
}
\seealso{
\code{\link{compromise_X3}} for an analogous function for the
compromise effect.
}
