% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T_1969_data.R
\docType{data}
\name{T_1969_counts}
\alias{T_1969_counts}
\title{Choice count data from the Tversky (1969) experiment}
\format{An object of class \code{array} of dimension 8 x 31 x 5.}
\usage{
T_1969_counts
}
\description{
An 8 by 31 by 5 array of count data.
Element \eqn{i,j,k} gives the number of times subject \eqn{i} chose object
\eqn{k} when presented with choice set \eqn{j}.
The choice set index \eqn{j \in \{1,...,7\}} encodes a non-empty subset of the
universe of the three choice objects \eqn{x}, \eqn{y}, and \eqn{z}, numbered 1, 2, 3.
Whenever \eqn{k} is not an element of \eqn{j}, the value is \code{NA}.
}
\seealso{
{\code{\link{RanCh}}}, under \code{Datasets} for a description of the experiment.
\code{\link{set_index}} describes how the choice set index encodes the choice set.
}
\keyword{datasets}
