% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YG_data.R
\docType{data}
\name{YG_demographics}
\alias{YG_demographics}
\title{Demographic information for subjects}
\format{A data frame with demographic information on subjects
\describe{
\item{\code{sex}}{Sex of subject}
\item{\code{educ}}{Educational attainment by subject}
\item{\code{region}}{Region of subject's residence in US}
\item{\code{race}}{Race of subject}
\item{\code{age_range}}{Age range of subject}
}}
\usage{
YG_demographics
}
\description{
Demographic information for subjects
}
\keyword{datasets}
