% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YG_data.R
\docType{data}
\name{YG_raw}
\alias{YG_raw}
\title{YouGov Experiment data}
\format{A data frame with 17 variables:
\describe{
\item{\code{design}}{}
\item{\code{card}}{}
\item{\code{domain}}{}
\item{\code{combo}}{}
\item{\code{perm}}{}
\item{\code{choiceset}}{Choice set as a character string}
\item{\code{option_1}}{Object presented in first position: 1, 2, 3 or 4}
\item{\code{option_2}}{Object presented in second position}
\item{\code{option_3}}{Object presented in third position}
\item{\code{option_4}}{Object presented in fourth position}
\item{\code{response}}{Object chosen: 1, 2, 3 or 4}
\item{\code{order}}{}
\item{\code{gender}}{Sex of respondant: 1 for male, 2 for female}
\item{\code{educ}}{Education of respondant: 1 for No high school, 2 for High school graduate,
3 for Some college, 4 for 2-year college, 5 for 4-year college, 6 for post-graduate}
\item{\code{region}}{Region of respondant: 1 for northeast, 2 for midwest, 3 for south, 4 for west}
\item{\code{race}}{Race of respondant: 1 for White, 2 for Black, 3 for Hispanic, 4 for Asian,
5 for Native American, 6 for Mixed, 7 for Other, 8 for Middle Eastern}
\item{\code{age_cross}}{Age category of respondant: 1 for 18-34, 2 for 35-54, 3 for 55 and over}
}}
\usage{
YG_raw
}
\description{
Raw trial-by-trial data from YouGov experiment
}
\keyword{datasets}
