% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YG_data.R
\docType{data}
\name{YG_trials}
\alias{YG_trials}
\title{Record of all choice trials}
\format{A data frame with 14 variables
\describe{
\item{\code{subj}}{Subject identifier}
\item{\code{domain}}{Factor indicating choice domain}
\item{\code{trial}}{Trial identifier (gives the order in which a subject sees choice sets)}
\item{\code{subs}}{Factor indicating the choice subset presented: 'ab', 'cde', etc.}
\item{\code{choice}}{Factor indicating the choice made: 'a', 'b', 'c' or 'd'}
\item{\code{subs_conf}}{Subset configuration, the order objects appear on the screen}
\item{\code{subs_bin}}{Code for subset where digits of binary representation indicate object membership}
\item{\code{choice_int}}{Integer code for chosen object}
\item{\code{ab}}{Revealed preference indicator: 1 for a revealed preferred to b, -1 for b revealed preferred to a, 0 otherwise}
}}
\usage{
YG_trials
}
\description{
Record of all choice trials
}
\keyword{datasets}
