% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HPD.R
\name{dDir}
\alias{dDir}
\title{Dirichlet density}
\usage{
dDir(p, alpha, log = TRUE)
}
\arguments{
\item{p}{vector of probabilities on the regular simplex}

\item{alpha}{vector of Dirichlet parameters}

\item{log}{logical; if TRUE, the log density is returned}
}
\value{
density or log density value
}
\description{
\code{dDir} computes the Dirichlet density at a point \code{p} in the
regular simplex, for a vector \code{alpha} of Dirichlet parameters.
}
