% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{log_L_DCE_Dir_mult}
\alias{log_L_DCE_Dir_mult}
\title{Log likelihood for DCE, multiple Dirichlet-multinomial model}
\usage{
log_L_DCE_Dir_mult(Alpha, N, log = TRUE)
}
\arguments{
\item{Alpha}{matrix of Dirichlet parameters, each row giving the Dirichlet
distribution of the corresponding row of a random choice structure.}

\item{N}{count matrix with the same dimensions as \code{Alpha}, pertaining to the same
universe of objects.}

\item{log}{logical; if TRUE, return the log Bayes factor}
}
\value{
Likelihood or log likelihood value
}
\description{
\code{log_L_DCE_Dir_mult} computes the marginal likelihood for a model
where choice count vectors are independent multinomial across choice sets
and choice probability vectors are independent Dirichlet across choice sets.
}
\examples{
Alpha = RCS_scalar_alpha_prior(2.0, 3)
N = T_1972_counts['Dots', 6,,]
log_L_DCE_Dir_mult(Alpha, N)
}
\seealso{
\code{\link{sim_DCE_Dir_mult}}, which simulates a count matrix under
this model, given the total number of trials for each choice subset.
}
