% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{log_L_Dir_mult}
\alias{log_L_Dir_mult}
\title{Log likelihood, Dirichlet-multinomial model}
\usage{
log_L_Dir_mult(alpha, n, log = TRUE)
}
\arguments{
\item{alpha}{vector of Dirichlet parameters}

\item{n}{vector of multinomial counts}

\item{log}{logical; if \code{TRUE}, return the log likelihood;
if \code{FALSE}, the likelihood.
\code{log=FALSE} is usually not recommendend, as underflow is likely.}
}
\value{
Marginal likelihood or log marginal likelihood
}
\description{
\code{log_L_Dir_mult} computes the log marginal likelihood for a multinomial
data generating process and a Dirichlet prior over choice probabilities.
}
