% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginalize.R
\name{marginalize}
\alias{marginalize}
\title{Routines for simple manipulations of count matrices and random choice structures.}
\usage{
marginalize(input_N, objects)
}
\arguments{
\item{input_N}{A count matrix}

\item{objects}{A vector of objects to retain}
}
\value{
A count matrix
}
\description{
Marginalize a count matrix or random choice structure
}
\details{
This function takes as input a count matrix or random choice structure on a
universe of objects and returns a marginalization of it to a universe that is a
subset of the original universe.
}
\examples{
N_bce <- marginalize(PC_counts['Beer',,], c(2, 3, 5)) # Marginalize data first
P_bce_1 <- proportions(N_bce)                         # then compute proportions
P <- proportions(PC_counts['Beer',,])   # Compute proportions first
P_bce_2 <- marginalize(P, c(2, 3, 5))   # then marginalize. Gives same result.
}
