% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PC_data.R
\docType{data}
\name{PC_counts}
\alias{PC_counts}
\title{Array of choice count data from the Population Choice experiment}
\format{An object of class \code{array} of dimension 32 x 31 x 5.}
\usage{
PC_counts
}
\description{
A three dimensional array, 32 by 31 by 5, of count data.
Element \eqn{i,j,k} gives the number of subjects that chose object \eqn{k} when
presented with choice set \eqn{j} of domain \eqn{i}.
The choice set index \eqn{j\in\{1,...,31\}} encodes a non-empty subset of the universe
of the five choice objects \eqn{a}, \eqn{b}, \eqn{c}, \eqn{d} and \eqn{e}, numbered 1, 2, 3, 4, 5.
Each digit (or bit) in the binary representation of \eqn{j} is an inclusion indicator: object k
is in the set if and only if the \eqn{k}'th digit from the right is 1.
For example, the set \eqn{\{b, c, d, e\}} is encoded as binary 11110 (decimal 30)
Whenever \eqn{k} is not an element of \eqn{j}, the value is \code{NA}.
}
\seealso{
{\code{\link{RanCh}}}, under \code{Datasets} for a description of the experiment.
Other data objects for the experiment include
\code{\link{PC_trials}}, a table of choice trial data,
\code{\link{PC_demographics}}, a table of demographic information, and
\code{\link{PC_raw}}, a table of raw data.
}
\keyword{datasets}
