% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PC_data.R
\docType{data}
\name{PC_counts}
\alias{PC_counts}
\title{Choice counts in Population Choice experiment}
\format{An object of class \code{array} of dimension 32 x 31 x 5.}
\usage{
PC_counts
}
\description{
A three dimensional array, 32 by 31 by 5, of count data.
Element i,j,k gives the number of subjects that chose object k when
presented with choice set j of domain i.
The choice set index j=1,...,31 encodes a non-empty subset of the universe
of the five choice objects a, b, c, d and e, numbered 1, 2, 3, 4, 5.
Each digit (or bit) in the binary representation of j is an inclusion indicator: object k
is in the set if and only if the k'th digit from the right is 1.
For example, the set with b, c, d and e (but not a) is encoded as binary 11110 (decimal 30)
Whenever k is not an element of j, the value is NA.
}
\seealso{
{\code{\link{RanCh}}}, under \code{Datasets} for a description of the experiment.
Other data objects for the experiment include
\code{\link{PC_trials}} for trial-by-trial data,
\code{\link{PC_demographics}} for demographic information and
\code{\link{PC_raw}} for the original raw data.
}
\keyword{datasets}
