% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PC_data.R
\docType{data}
\name{PC_counts}
\alias{PC_counts}
\title{Counts}
\format{An object of class \code{array} of dimension 32 x 31 x 5.}
\usage{
PC_counts
}
\description{
A 32x26x5 matrix with count data.
}
\keyword{datasets}
