% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PC_data.R
\docType{data}
\name{PC_demographics}
\alias{PC_demographics}
\title{Demographic information for population choice experiment}
\format{A data frame with demographic information on subjects
\describe{
\item{\code{sex}}{Sex of subject: 1 for male, 2 for female}
\item{\code{age}}{Age of subject in years}
\item{\code{location}}{Province or territory in Canada, 1=Alberta, 2=British Columbia,
3=Manitoba, 4=New Brunswick, 5=Newfoundland/Labrador, 6=Northwest Territories,
7=Nova Scotia, 8=Ontario, 9=Prince Edward Island, 10=Quebec, 11=Saskatchewan, 12=Yukon}
}}
\usage{
PC_demographics
}
\description{
Demographic information for population choice experiment
}
\seealso{
{\code{\link{RanCh}}}, under \code{Datasets} for a description of the experiment.
Other data objects for the experiment include
\code{\link{PC_trials}} for trial-by-trial data,
\code{\link{PC_counts}} for choice count data and
\code{\link{PC_raw}} for the original raw data.
}
\keyword{datasets}
