% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PC_data.R
\docType{data}
\name{PC_demographics}
\alias{PC_demographics}
\title{Demographic information for subjects}
\format{A data frame with demographic information on subjects
\describe{
\item{\code{sex}}{Sex of subject }
\item{\code{age}}{Age of subject in years}
\item{\code{location}}{Province or territory in Canada}
}}
\usage{
PC_demographics
}
\description{
Demographic information for subjects
}
\keyword{datasets}
