% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PC_data.R
\docType{data}
\name{PC_trials}
\alias{PC_trials}
\title{Table of choice trials, population choice experiment}
\format{A tibble with 20 variables
\describe{
\item{\code{domain}}{factor, name of choice domain}
\item{\code{subject}}{subject identifier}
\item{\code{trial}}{trial identifier (gives the order in which a subject sees choice sets)}
\item{\code{duration}}{duration of trial in seconds}
\item{\code{set}}{factor, name of choice set presented: 'ab', 'cde', etc., with objects
in alphabetical order}
\item{\code{choice}}{factor, choice made by subject: 'a', 'b', 'c', 'd' or 'e'}
\item{\code{set_perm}}{factor, order of presentation of objects on screen, left to right}
\item{\code{set_card}}{Integer, cardinality of choice set (i.e. number of available options)}
\item{\code{set_bin}}{Binary representation of choice set (binary digits indicate object membership in choice set)}
\item{\code{choice_int}}{Integer code for chosen object: a=1, b=2, ..., e=5}
\item{\code{ab}, \code{ac}, \ldots, \code{de}}{revealed preference indicator: taking column ab as an example,
value is 1 if a is revealed preferred to b, -1 if b is revealed preferred to a, 0 otherwise.}
}}
\usage{
PC_trials
}
\description{
Table of choice trials, population choice experiment
}
\seealso{
\code{\link{RanCh}}, under \code{Datasets} for a description of the experiment.
Other data objects for the experiment include
\code{\link{PC_counts}} for choice count data,
\code{\link{PC_demographics}} for demographic information and
\code{\link{PC_raw}} for the original raw data.
}
\keyword{datasets}
