% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HPD.R
\name{plot_HD_Dir3}
\alias{plot_HD_Dir3}
\title{Plot highest density region for a third order Dirichlet distribution}
\usage{
plot_HD_Dir3(A, HD_probability, selection)
}
\arguments{
\item{HD_probability}{probability of highest density region}

\item{alpha}{vector of Dirichlet parameters}
}
\description{
This function plots the Dirichlet highest density region in barycentric
coordinates.
}
\examples{
plot_HD_Dir3(A, 0.95, c(1,2,3))
}
