% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RanCh.R
\docType{package}
\name{RanCh}
\alias{RanCh}
\alias{RanCh-package}
\title{RanCh: A package for abstract discrete Random Choice}
\description{
The RanCh package provides data, graphical tools and inference tools for
abstract discrete random choice analysis.
}
\section{Data sets}{
NA
}

