% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{RCS_vector_alpha_prior}
\alias{RCS_vector_alpha_prior}
\title{Vector alpha prior for a RCS}
\usage{
RCS_vector_alpha_prior(alpha, v)
}
\arguments{
\item{alpha}{scalar, prior precision parameter}

\item{v}{vector, object weights for the vector Dirichlet prior.}
}
\value{
a matrix of Dirichlet parameters for the vector Dirichlet prior.
}
\description{
\code{RCS_vector_alpha_prior} computes a matrix of Dirichlet parameters
for a vector Dirichlet prior for a random choice structure.
}
\examples{
Alpha <- RCS_vector_alpha_prior(5.2, c(2.5, 1.2, 3.3))
P <- sim_RCS_Dirichlet(10, Alpha)
}
