% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDS_2011_data.R
\docType{data}
\name{RDS_2011_counts}
\alias{RDS_2011_counts}
\title{Choice count data from the Regenwetter, Dana and Davis-Stober (2011) experiment}
\format{
An object of class \code{array} of dimension 3 x 18 x 31 x 5.
}
\usage{
RDS_2011_counts
}
\description{
A 3 by 18 by 31 by 5 array of count data.
Element \eqn{h,i,j,k} gives the number of times subject \eqn{i} chose object
\eqn{k} when presented with choice set \eqn{j} of domain \eqn{h}.
Domain names are "Cash 1", "Cash 2" and "Noncash", explained in the paper.
The choice set index \eqn{j \in \{1,...,31\}} encodes a non-empty subset of the
universe of the four choice objects a, b, c, d, and e, numbered 1, 2, 3, 4, 5.
Whenever \eqn{k} is not an element of \eqn{j}, the value is \code{NA}.
}
\seealso{
{\code{\link{RanCh}}}, under \code{Datasets} for a description of the experiment.
\code{\link{set_index}} describes how the choice set index encodes the choice set.
}
\keyword{datasets}
