% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YG_data.R
\docType{data}
\name{YG_counts}
\alias{YG_counts}
\title{Counts}
\format{An object of class \code{array} of dimension 16 x 2 x 15 x 4.}
\usage{
YG_counts
}
\description{
A 3x16x15x4 matrix with count data.
}
\keyword{datasets}
