% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YG_data.R
\docType{data}
\name{YG_demographics}
\alias{YG_demographics}
\title{Table of demographic information from the YouGov experiment}
\format{A data frame with demographic information on subjects
\describe{
\item{\code{sex}}{Sex of subject}
\item{\code{educ}}{Educational attainment by subject}
\item{\code{region}}{Region of subject's residence in US}
\item{\code{race}}{Race of subject}
\item{\code{age_range}}{Age range of subject}
}}
\usage{
YG_demographics
}
\description{
Table of demographic information from the YouGov experiment
}
\seealso{
{\code{\link{RanCh}}}, under \code{Datasets} for a description of the experiment.
Other data objects for the experiment include
\code{\link{YG_trials}}, a table of choice trial data,
\code{\link{YG_counts}}, a matrix of choice count data, and
\code{\link{YG_raw}}, a table of raw data.
}
\keyword{datasets}
